requirements.txt
flask>=3.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
yfinance>=0.2.40
lxml>=5.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
reportlab>=4.0.0
gunicorn>=21.0.0
gevent>=24.0.0
